pyproject.toml
[project]
name = "picnic-shopping-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.76.0",
    "python-picnic-api2",
    "python-telegram-bot>=22.5",
]
